Continued fixes for League of Legends modding (Issue #1).
** THIS SOLUTION INVOLVES AN INTENTIONAL MEMORY LEAK! Explained Below **

Dargon now seems to go in-game with a 100% success rate. However, this
solution involves allocating a dummy buffer (which is never freed!) for
overlapped IO. Furthermore, I've stopped tracking the file pointer
manually, opting to follow the old dargon's style of maintaining a dummy
file handle.

By the Windows API specification and this document:
   https://support.microsoft.com/en-us/kb/156932

It seems clear that an asynchronous ReadFile call (that is, one with an
OVERLAPPED structure) may, by specification, complete synchronously.
However, it seems like League of Legends does not like synchronous I/O
completion and instead starts another read, which is unfortunate.