Fix non-fatal vfm read bug where we'd ask sectors for too much data.
As a bonus, this commit simplifies the math needed to calculate source
sector and destination buffer offsets.

This wasn't the root cause of any memory-stomping issues. When we read
from a vfm, we calculate bytes_read which equals the maximum data we are
supposed to be able to read from the vfm. This same value is returned as
the number of bytes read.

When determining how much data to read for the last file_sector (see:
sectors_to_read), we were clamping the copy_length based on the buffer's
total size, rather than the amount of data we wanted to read. This could
lead to reading more data than necessary at the end (but still within)
the user's buffer. The user application was not affected because the
returned bytes_read does not indicate the existence of this data.