Remove codepath that placed raf overrides in overridden dat file space.
When generating a virtual file map for an overridden dat file, we had
two code paths: one that appends to the end of the virtual dat file and
another that inserts in the old raf-entry's spot. As there's no tangible
cost to appending to the end of the file (we're using vfm) it makes
little sense to have two code-paths, especially since with the mods
I'm testing, the 'optimize' case was rarely being run.